// web/src/api/generated/model/priceListDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PriceListLineInput } from './priceListLineInput';

export interface PriceListDto {
  id: string;
  nameAr: string;
  nameEn: string;
  currencyCode: string;
  isActive: boolean;
  lines: PriceListLineInput[];
  inUse: boolean;
}