// web/src/api/generated/model/priceListInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PriceListLineInput } from './priceListLineInput';

export interface PriceListInput {
  nameAr: string;
  nameEn: string;
  /** @nullable */
  currencyCode: string | null;
  lines: PriceListLineInput[];
}